+++ src/py42/_internal/archive_access.py
@@ -1,7 +1,9 @@
 import posixpath
 import time
+import json
 from collections import namedtuple
 
+from py42.clients import BaseClient
 from py42.sdk.exceptions import Py42ArchiveFileNotFoundError
 
 FileSelection = namedtuple(u"FileSelection", u"path_set, num_files, num_dirs, size")
@@ -10,6 +12,13 @@ FileSelection = namedtuple(u"FileSelection", u"path_set, num_files, num_dirs, si
 class FileType(object):
     DIRECTORY = u"directory"
     FILE = u"file"
+
+
+class ArchiveAccessClient(BaseClient):
+    def get_data_key_token(self, device_guid):
+        uri = u"/api/DataKeyToken"
+        data = {u"computerGuid": device_guid}
+        return self._session.post(uri, data=json.dumps(data))
 
 
 class ArchiveAccessorManager(object):
@@ -27,8 +36,6 @@ class ArchiveAccessorManager(object):
         return ArchiveAccessor(device_guid, session_id, client.archive, restore_job_manager)
 
     def _get_data_key_token(self, device_guid):
-        test = self._archive_client.get_data_key_token(device_guid)
-        test2 = test[u"dataKeyToken"]
         return self._archive_client.get_data_key_token(device_guid)[u"dataKeyToken"]
 
     @staticmethod